# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HealApp API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Register a new user with phone number and password. A verification code will be sent via SMS.",
                "operationId": "15ab81fc91b8e90f74ea030fc44604ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password",
                                    "password_confirmation",
                                    "account_type"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "User's first name (optional)",
                                        "type": "string",
                                        "example": "John",
                                        "nullable": true
                                    },
                                    "last_name": {
                                        "description": "User's last name (optional)",
                                        "type": "string",
                                        "example": "Doe",
                                        "nullable": true
                                    },
                                    "middle_name": {
                                        "description": "User's middle name (optional)",
                                        "type": "string",
                                        "example": "Michael",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "description": "User's phone number (must be unique)",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "password": {
                                        "description": "User's password (minimum 6 characters)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "description": "Password confirmation",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "account_type": {
                                        "description": "Account type: client, specialist, pansionat, or agency",
                                        "type": "string",
                                        "enum": [
                                            "client",
                                            "specialist",
                                            "pansionat",
                                            "agency"
                                        ],
                                        "example": "client"
                                    },
                                    "organization_name": {
                                        "description": "Organization name (optional, used for pansionat or agency)",
                                        "type": "string",
                                        "example": "Health Care Center",
                                        "nullable": true
                                    },
                                    "address": {
                                        "description": "Address (optional, used for organizations)",
                                        "type": "string",
                                        "example": "123 Main St, City",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration successful. SMS sent.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "SMS sent"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "1234567890"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Revoke the current access token for the authenticated user.",
                "operationId": "95268babc80be6aea5dffe685834b88c",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}